MoviesPresenter creation of ui package and moving view model and state there. Also renaming view package to components
