+++ app/src/main/java/com/tomerpacific/moviepresenter/MainActivity.kt
@@ -10,9 +10,9 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.tomerpacific.moviepresenter.model.MainViewModel
+import com.tomerpacific.moviepresenter.ui.main.MainViewModel
 import com.tomerpacific.moviepresenter.ui.theme.MoviePresenterTheme
-import com.tomerpacific.moviepresenter.ui.view.MovieView
+import com.tomerpacific.moviepresenter.ui.components.MovieView
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.lifecycle.ViewModelProvider
@@ -20,8 +20,8 @@ import com.tomerpacific.moviepresenter.domain.repository.MovieRepositoryImpl
 import com.tomerpacific.moviepresenter.domain.usecase.GetMoviePosterUseCase
 import com.tomerpacific.moviepresenter.domain.usecase.GetMoviePostersUseCase
 import com.tomerpacific.moviepresenter.domain.usecase.GetMoviesUseCase
-import com.tomerpacific.moviepresenter.model.MainViewModelFactory
-import com.tomerpacific.moviepresenter.ui.view.MovieList
+import com.tomerpacific.moviepresenter.ui.main.MainViewModelFactory
+import com.tomerpacific.moviepresenter.ui.components.MovieList
 
 class MainActivity : ComponentActivity() {
