+++ app/src/main/java/com/tomerpacific/moviepresenter/ui/components/CircularProgressBarIndicator.kt
@@ -1,4 +1,4 @@
-package com.tomerpacific.moviepresenter.ui.view
+package com.tomerpacific.moviepresenter.ui.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
