+++ app/src/main/java/com/tomerpacific/moviepresenter/ui/components/MovieCard.kt
@@ -1,4 +1,4 @@
-package com.tomerpacific.moviepresenter.ui.view
+package com.tomerpacific.moviepresenter.ui.components
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
@@ -17,7 +17,7 @@ import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.tomerpacific.moviepresenter.model.MainViewModel
+import com.tomerpacific.moviepresenter.ui.main.MainViewModel
 import com.tomerpacific.moviepresenter.domain.model.MovieModel
 
 @Composable
