+++ app/src/main/java/com/tomerpacific/moviepresenter/ui/main/MainViewModel.kt
@@ -1,4 +1,4 @@
-package com.tomerpacific.moviepresenter.model
+package com.tomerpacific.moviepresenter.ui.main
 
 import android.app.Application
 import android.graphics.Bitmap
